+++ src/licensing/crud/hierarchy_provider.py
@@ -1,19 +1,22 @@
 from functools import lru_cache
+from typing import Dict
 
 from fastapi import status as http_status, HTTPException
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from licensing.hierarchy_provider_client import get_user_memberships_from_provider
+from licensing.client.hierarchy_provider_client import hierarchy_provider_memberships_url
+from licensing.http_client import http_get
 from licensing.model import hierarchy_provider as model
+from licensing.utils import async_measure_time
 
 
 @lru_cache()
-async def get_hierarchy_provider(session: AsyncSession, url: str) -> model.HierarchyProvider:
+async def find_hierarchy_provider(session: AsyncSession, url: str) -> model.HierarchyProvider:
     """
-    gets a hierarchy provider with a given url or nothing
+    finds a hierarchy provider with a given url or raises an HTTPException, if not found
     """
-    return (
+    provider = (
         await session.execute(
             statement=select(
                 model.HierarchyProvider
@@ -22,14 +25,6 @@ async def get_hierarchy_provider(session: AsyncSession, url: str) -> model.Hiera
             )
         )
     ).scalar_one_or_none()
-
-
-@lru_cache()
-async def find_hierarchy_provider(session: AsyncSession, url: str) -> model.HierarchyProvider:
-    """
-    finds a hierarchy provider with a given url or raises an HTTPException, if not found
-    """
-    provider = await get_hierarchy_provider(session, url)
     if not provider:
         raise HTTPException(
             status_code=http_status.HTTP_400_BAD_REQUEST,
@@ -38,16 +33,20 @@ async def find_hierarchy_provider(session: AsyncSession, url: str) -> model.Hier
     return provider
 
 
-async def get_user_memberships(session: AsyncSession, url: str, user_eid) -> (model.HierarchyProvider, set):
+@async_measure_time
+async def get_user_memberships(
+        session: AsyncSession, url: str, user_eid: str
+) -> (model.HierarchyProvider, Dict[(str, int)]):
     """
     gets the membership list for a user by requesting the hierarchy provider.
     catches any exception raised from the request and 'translate' it to some 500
     http error.
-    returns a tuple (the registered hierarchy provider object, the hierarchy itself as a set!!)
+    returns a tuple (the registered hierarchy provider object, the memberships as a dict (key=entity EID)!)
     """
-    hp = await find_hierarchy_provider(session, url)  # raises exception, if not found
+    hierarchy_provider = await find_hierarchy_provider(session, url)
     try:
-        return hp, set(await get_user_memberships_from_provider(url, user_eid))
+        memberships_raw = await http_get(hierarchy_provider_memberships_url(f"{url}/", user_eid))
+        return hierarchy_provider, {m["eid"]: (m["type"], m['level']) for m in memberships_raw}
     except Exception as e:
         raise HTTPException(
             status_code=http_status.HTTP_500_INTERNAL_SERVER_ERROR,
